requirements.txt
yfinance>=0.2.36
gspread>=6.0.0
google-auth>=2.27.0
